--refactor:: setup components page to display how our components are being used
