+++ src/pages/app/Overview.tsx
@@ -1,0 +1,8 @@
+
+export const Overview = ({}) => {
+ return ( 
+  <>
+   <h1>Overview page</h1>
+  </>
+ );
+}
