+++ src/pages/index.ts
@@ -1,0 +1,4 @@
+export * from "./auth/Login";
+export * from "./auth/Signup";
+export * from "./app/Overview";
+export * from "./components/Component";
